chore(BA-2964): Move `EndpointLifecycle` eum to common package (#6637)
Backported-from: main (25.17)
Backported-to: 25.15
Backport-of: 6637
