+++ changes/6637.misc.md
@@ -1,0 +1,1 @@
+Move `EndpointLifecycle` enum to a shared common package for improved reusability
