+++ src/ai/backend/common/data/endpoint/types.py
@@ -1,4 +1,5 @@
-from enum import StrEnum
+from enum import Enum, StrEnum
+from functools import lru_cache
 
 
 class EndpointStatus(StrEnum):
@@ -9,3 +10,27 @@ class EndpointStatus(StrEnum):
     DESTROYING = "DESTROYING"
     DESTROYED = "DESTROYED"
     DEGRADED = "DEGRADED"
+
+
+class EndpointLifecycle(Enum):
+    PENDING = "pending"
+    CREATED = "created"  # Deprecated, use READY instead
+    SCALING = "scaling"
+    READY = "ready"
+    DESTROYING = "destroying"
+    DESTROYED = "destroyed"
+
+    @classmethod
+    @lru_cache(maxsize=1)
+    def active_states(cls) -> set["EndpointLifecycle"]:
+        return {cls.PENDING, cls.CREATED, cls.SCALING, cls.READY}
+
+    @classmethod
+    @lru_cache(maxsize=1)
+    def need_scaling_states(cls) -> set["EndpointLifecycle"]:
+        return {cls.CREATED, cls.READY}
+
+    @classmethod
+    @lru_cache(maxsize=1)
+    def inactive_states(cls) -> set["EndpointLifecycle"]:
+        return {cls.DESTROYING, cls.DESTROYED}
